authentication/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asarPy==1.0.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
fastapi==0.110.3
h11==0.14.0
idna==3.7
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
